Move VSCleanup and test code to volumesnapshot package
