+++ pkg/volumesnapshot/volumesnapshot_cleanup_policy_test.go
@@ -1,4 +1,4 @@
-package util
+package volumesnapshot
 
 import (
 	"fmt"
